feat: 添加样式注册功能和吉他谱文件检查函数
